Ensure custody of timezones from query to results, add debugging info
This commit makes several changes intended to fix timezone related
issues and expose more information when there are problems. For
databases that support report timezones, this commit fixes an issue
where the JVM's timezone was used to interpret datetimes that didn't
include a timezone. The report timezone should be used instead. This
commit also adds a step to the sync process to detect and store the
databases timezone, which will be helpful in troubleshooting timezone
issues with a new timezone debugging page (to be added later).